CRUD on books

+++ components/BookCard.js
@@ -9,7 +9,9 @@ function BookCard({ bookObj, onUpdate }) {
   // FOR DELETE, WE NEED TO REMOVE THE BOOK AND HAVE THE VIEW RERENDER,
   // SO WE PASS THE FUNCTION FROM THE PARENT THAT GETS THE BOOKS
   const deleteThisBook = () => {
-    deleteBook(bookObj.firebaseKey).then(() => onUpdate());
+    if (window.confirm(`Delete ${bookObj.title}?`)) {
+      deleteBook(bookObj.firebaseKey).then(() => onUpdate());
+    }
   };
 
   return (
@@ -17,16 +19,16 @@ function BookCard({ bookObj, onUpdate }) {
       <Card.Img variant="top" src={bookObj.image} alt={bookObj.title} style={{ height: '400px' }} />
       <Card.Body>
         <Card.Title>{bookObj.title}</Card.Title>
-        <p className="card-text bold">{bookObj.sale ? bookObj.price : bookObj.price}</p>
+        <p className="card-text bold">{bookObj.sale && <span>SALE<br /></span> } ${bookObj.price}</p>
         {/* DYNAMIC LINK TO VIEW THE BOOK DETAILS  */}
         <Link href={`/book/${bookObj.firebaseKey}`} passHref>
-          <Button variant="primary">VIEW</Button>
+          <Button variant="primary" className="m-2">VIEW</Button>
         </Link>
         {/* DYNAMIC LINK TO EDIT THE BOOK DETAILS  */}
         <Link href={`/book/edit/${bookObj.firebaseKey}`} passHref>
           <Button variant="info">EDIT</Button>
         </Link>
-        <Button variant="danger" onClick={deleteThisBook}>
+        <Button variant="danger" onClick={deleteThisBook} className="m-2">
           DELETE
         </Button>
       </Card.Body>
